auth-bot-fastapi/requirements.txt
python-telegram-bot
fastapi
uvicorn
redis
PyJWT
requests
aiogram==2.25.1
asyncio